scheduler/requirements.txt
fastapi[standard]==0.115.6
pydantic==2.10.4
python-dotenv==1.0.1
motor==3.6.0
firebase-admin==6.6.0
APScheduler==3.11.0
pytz==2024.2